Refactor sign on methods component
